Added significantly improved version of getip.sh script
The script now supports resolving both IPv4 and IPv6 addresses. In addition it
now uses getent so that entries in /etc/hosts are also read, and preferred over
those found from DNS. Note that it's not recommended to bind a localhost address
to the FQDN in /etc/hosts for the node that runs this script (=puppetmaster); if
this is done a localhost address might be returned for that node, resulting in
non-functional Puppet resources.

Signed-off-by: Samuli Seppänen <samuli@openvpn.net>