Install all validation scripts by default (v0.3.2)
Previously the values of $some_check parameters determined three things:

- Is the script itself present?
- Is the script invocation in the (interactive) pvalidate.sh wrapper?
- Does the script run from cron?

This was suboptimal, because pvalidate.sh with the full battery of test is very
useful during development, but having cron email tons of puppet-lint errors and
warnings for no good reason is not. Now all scripts are installed and added to
pvalidate.sh automatically and their presence in crontab is determined by the
class parameters.

Signed-off-by: Samuli Seppänen <samuli@openvpn.net>