Integrate extended ACL management
Extended ACLs are useful when several admins have to be able to write files to
and "git commit" in /etc/puppetlabs. Previously the ACLs had to be set manually
using $::setfacl::targets hash, but now it is usually enough to set
$::puppetmaster::manage_acls to true. Note that this commit changes the
permissions of some files from '0644' to '0654' to avoid the File resource from
"fixing" the permissions on every run when extended ACLs are in use.

Signed-off-by: Samuli Seppänen <samuli.seppanen@gmail.com>