Use $title as the default to IPv4 and IPv6 packet filter rules
This change allows one to create allow rules based on the resource
(ipv4_allow/ipv6_allow) $title. While not the best approach security-vise, this
is highly useful when a node connect to the Puppetmaster through the public
Internet and its IP changes often, but the A/AAAA records are kept in sync with
the IP. Note that this change is invasive and can cause Puppet run failures due
to significant refactoring.

Signed-off-by: Samuli Seppänen <samuli@openvpn.net>